fix : remove old card and updated new cards on ui
